+++ release_notes.md
@@ -6,10 +6,20 @@
   - ability to see what if any target is currently being edited using the `active_target_id` property
   - added custom key actions and overlay layer support
   - added pre and post insert hooks
+  - generalized implementation to support multiple target types
+    - added support for `CTkTextBoxEx` targets
 - `el.datastore.SavableModel`
   - added ClassVar `model_dump_default_options` to  to specify default options used for saving files (like indentation).
+- `el.widgets.ctkex.CTkEntryEx`
+  - `select_all()` and Ctrl+A selection by default
 
 ## New Features
 
 - added `el.base` module
   - `filter_kwargs` helper to only pass non-None kwargs on another function
+- added `el.widgets.ctkex.CTkTextBoxEx`
+  - touchscreen mode
+  - `select_all()` and Ctrl+A selection by default
+  - ability to exclude corners from radius
+  - ability to change background corner colors
+  - ability to control width/height rounding